split-up lexer and style
